moved class to analytics dir
