+++ lib/analytics/CTAnalyticsManager.dart
