+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:corona_trace/CTAnalyticsManager.dart';
+import 'package:corona_trace/analytics/CTAnalyticsManager.dart';
 import 'package:corona_trace/network/api_repository.dart';
 import 'package:corona_trace/ui/notifications/notification_list_screen.dart';
 import 'package:corona_trace/ui/screens/onboarding.dart';
